New_Zaawansowane_Programowanie/requirements.txt
aniso8601==9.0.1
antiorm==1.2.1
asgiref==3.5.2
auth==0.5.3
bcrypt==4.0.1
blinker==1.5
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
db==0.1.1
Django==4.1.3
dnspython==2.2.1
email-validator==1.3.0
eventlet==0.33.2
falcon==3.1.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
flask-swagger-ui==4.11.1
Flask-WTF==1.0.1
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
image==1.5.33
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.2
numpy==1.23.5
Pillow==9.3.0
pymongo==4.3.3
pytz==2022.6
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.44
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12
Werkzeug==2.2.2
WTForms==3.0.1
